- remove safe

+++ src/document/document.service.ts
@@ -1,4 +1,4 @@
-import { ConsoleLogger, Injectable } from '@nestjs/common';
+import { ConsoleLogger, Injectable, NotFoundException } from '@nestjs/common';
 import { UpdateDocumentDto } from './dto/update-document.dto';
 import { IngestService } from '../ingest/ingest.service';
 import { VectorService } from '../ingest/vector/vector.service';
@@ -6,6 +6,7 @@ import { deleteFile } from './oss';
 import path from 'path';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateDocumentDto } from './dto/create-document.dto';
+import { project_documents } from '@prisma/client';
 
 @Injectable()
 export class DocumentService {
@@ -23,10 +24,22 @@ export class DocumentService {
     projectId?: string,
   ): Promise<void> {
     for (const file of files) {
+      let document: project_documents | null = null;
       try {
+        // Pre create document record with 'processing' status
+        document = await this.createDocument({
+          projectId: projectId as string,
+          name: file.originalname,
+          filePath: file.path,
+          mimeType: file.mimetype,
+          size: file.size,
+          status: 'processing',
+          userId: userId,
+        });
+
         const chunksCount = await this.ingestService.ingestDocument(
           file.path,
-          file.filename,
+          document.id,
           userId,
           projectId,
         );
@@ -37,15 +50,9 @@ export class DocumentService {
 
         // If ingestion successful (has chunks), save document record in DB
         if (chunksCount > 0) {
-          await this.createDocument({
-            projectId: projectId as string,
-            name: file.originalname,
-            filePath: file.path,
-            mimeType: file.mimetype,
-            size: file.size,
-            status: 'done',
-            userId: userId,
-          });
+          // update 'done' status
+          await this.updateDocumentStatusDone(document.id, 'done');
+
           this.logger.log(
             `📝 Document record created for: ${file.originalname}`,
           );
@@ -54,30 +61,43 @@ export class DocumentService {
         }
       } catch (error) {
         this.logger.error(`❌ Failed to ingest ${file.originalname}:`, error);
-        // Optionally save with error status
-        await this.createDocument({
-          projectId: projectId as string,
-          name: file.originalname,
-          filePath: file.path,
-          mimeType: file.mimetype,
-          size: file.size,
-          status: 'error',
-          userId: userId,
-        });
+        // Optionally update 'error' status
+        if (document) {
+          await this.updateDocumentStatusDone(document.id, 'error');
+        }
       }
     }
   }
   // -- REMOVE --
-  async removeDocument(fileId: string) {
-    // TODO: Check ownership or permissions
+  async removeDocument(fileId: string, userId: string) {
+    // Check doc exists
+    const document = await this.prisma.project_documents.findUnique({
+      where: { id: fileId },
+    });
+    if (!document) throw new NotFoundException('Document not found');
+    if (document.userId !== userId)
+      throw new NotFoundException('Document not found');
+
+    // Remove vectors
     await this.vectorService.removeVectorByFileId(fileId);
-    deleteFile(path.join('uploads/documents', fileId));
-    return fileId;
+
+    // Delete physical file
+    try {
+      deleteFile(path.join('uploads/documents', fileId));
+    } catch (error) {
+      console.error('⚠️ File delete error:', error);
+      throw new NotFoundException('Delete file uploads error');
+    }
+
+    // project_documents
+    return await this.prisma.project_documents.delete({
+      where: { id: fileId },
+    });
   }
 
   // -- CREATE DOCUMENT MAPPING --
   async createDocument(documentDto: CreateDocumentDto) {
-    await this.prisma.project_documents.create({
+    return await this.prisma.project_documents.create({
       data: {
         projectId: documentDto.projectId,
         name: documentDto.name,
@@ -118,6 +138,15 @@ export class DocumentService {
       where: { id: id },
       data: {
         name: updateDocumentDto.name,
+      },
+    });
+  }
+
+  async updateDocumentStatusDone(id: string, status: string) {
+    return await this.prisma.project_documents.update({
+      where: { id: id },
+      data: {
+        status: status,
       },
     });
   }
